move the CRE CLI to core/scripts and update docs
