+++ .gitignore
@@ -115,3 +115,8 @@ contracts/out/
 /lib/
 /build/bin/
 /build/lib/
+
+# artifacts of CRE environment creatio or CRE CLI usage
+core/scripts/cre/environment/*.yaml
+core/scripts/cre/environment/*.yml
+core/scripts/cre/environment/logs/
