+++ core/scripts/cre/environment/.gitignore
@@ -1,0 +1,21 @@
+# ctf binary artifacts
+compose/
+blockscout/
+
+# contract artifacts
+examples/contracts/abi
+examples/contracts/bin
+
+# CRE CLI binary
+cre_v*
+
+# crib-specific CL node configuration and secrets that will be used to override the default configuration
+crib-configs/
+cache/
+
+logs/
+
+# CRE CLI settings
+.cre.settings.yaml
+cre.yaml
+*.state.yaml
