+++ core/scripts/cre/environment/configs/single-don.toml
@@ -27,7 +27,7 @@
 
 [[nodesets]]
   nodes = 5
-  override_mode = "all"
+  override_mode = "each"
   http_port_range_start = 10100
   name = "workflow"
 
@@ -37,7 +37,21 @@
 
   [[nodesets.node_specs]]
     [nodesets.node_specs.node]
-      docker_ctx = "../../../../.."
+      docker_ctx = "../../../.."
+      docker_file = "plugins/chainlink.Dockerfile"
+      # image = "chainlink-tmp:latest"
+      # omitted on purpose, if not defined here it will be added by Go code if worker nodes from this DON have cron capability
+      # capabilities = ['./cron']
+      custom_ports = ["5002:5002"]
+      user_config_overrides = """
+      [Log]
+      Level = 'debug'
+      JSONConsole = true
+      """
+
+  [[nodesets.node_specs]]
+    [nodesets.node_specs.node]
+      docker_ctx = "../../../.."
       docker_file = "plugins/chainlink.Dockerfile"
       # image = "chainlink-tmp:latest"
       # omitted on purpose, if not defined here it will be added by Go code if worker nodes from this DON have cron capability
@@ -48,43 +62,39 @@
       JSONConsole = true
       """
 
-[[nodesets]]
-  nodes = 2
-  override_mode = "all"
-  http_port_range_start = 10200
-  name = "capabilities"
-
-  [nodesets.db]
-    image = "postgres:12.0"
-    port = 13100
-
   [[nodesets.node_specs]]
     [nodesets.node_specs.node]
-      docker_ctx = "../../../../.."
+      docker_ctx = "../../../.."
       docker_file = "plugins/chainlink.Dockerfile"
       # image = "chainlink-tmp:latest"
+      # omitted on purpose, if not defined here it will be added by Go code if worker nodes from this DON have cron capability
+      # capabilities = ['./cron']
       user_config_overrides = """
       [Log]
       Level = 'debug'
       JSONConsole = true
       """
 
-[[nodesets]]
-  nodes = 1
-  override_mode = "each"
-  http_port_range_start = 10300
-  name = "gateway"
-
-  [nodesets.db]
-    image = "postgres:12.0"
-    port = 13200
+  [[nodesets.node_specs]]
+    [nodesets.node_specs.node]
+      docker_ctx = "../../../.."
+      docker_file = "plugins/chainlink.Dockerfile"
+      # image = "chainlink-tmp:latest"
+      # omitted on purpose, if not defined here it will be added by Go code if worker nodes from this DON have cron capability
+      # capabilities = ['./cron']
+      user_config_overrides = """
+      [Log]
+      Level = 'debug'
+      JSONConsole = true
+      """
 
   [[nodesets.node_specs]]
     [nodesets.node_specs.node]
-      docker_ctx = "../../../../.."
+      docker_ctx = "../../../.."
       docker_file = "plugins/chainlink.Dockerfile"
       # image = "chainlink-tmp:latest"
-      custom_ports = ["5002:5002"]
+      # omitted on purpose, if not defined here it will be added by Go code if worker nodes from this DON have cron capability
+      # capabilities = ['./cron']
       user_config_overrides = """
       [Log]
       Level = 'debug'
