+++ core/scripts/cre/environment/examples/.tools-version
