+++ core/scripts/cre/environment/examples/contracts/Makefile
