+++ core/scripts/cre/environment/examples/contracts/permissionless_feeds_consumer/permissionless_feeds_consumer.go
