+++ core/scripts/cre/environment/examples/pkg/deploy/consumer.go
@@ -5,12 +5,12 @@ import (
 	"os"
 	"time"
 
-	"cosmossdk.io/errors"
 	"github.com/ethereum/go-ethereum/common"
+	"github.com/pkg/errors"
 
 	"github.com/smartcontractkit/chainlink-testing-framework/seth"
 
-	"github.com/smartcontractkit/chainlink/system-tests/tests/smoke/cre/cmd/examples/contracts/permissionless_feeds_consumer"
+	"github.com/smartcontractkit/chainlink/core/scripts/cre/environment/examples/contracts/permissionless_feeds_consumer"
 )
 
 func DeployPermissionlessFeedsConsumer(rpcUrl string) (*common.Address, error) {
