+++ core/scripts/cre/environment/examples/workflows/proof-of-reserve/workflow_data.json
