+++ system-tests/tests/smoke/cre/cmd/main.go
@@ -5,15 +5,11 @@ import (
 	"os"
 
 	"github.com/smartcontractkit/chainlink/system-tests/tests/smoke/cre/cmd/download"
-	"github.com/smartcontractkit/chainlink/system-tests/tests/smoke/cre/cmd/environment"
-	"github.com/smartcontractkit/chainlink/system-tests/tests/smoke/cre/cmd/examples"
 	"github.com/smartcontractkit/chainlink/system-tests/tests/smoke/cre/cmd/root"
 )
 
 func init() {
 	root.RootCmd.AddCommand(download.DownloadCmd)
-	root.RootCmd.AddCommand(environment.EnvironmentCmd)
-	root.RootCmd.AddCommand(examples.ExamplesCmd)
 }
 
 func main() {
